requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==5.1.1
django-filter==24.3
django-rest-swagger==2.2.0
django-tables2==2.7.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
